se separa por archivos todos los mapeadores
