+++ Mapeadores/Mapeadores.py
@@ -1,23 +1,15 @@
 from Entidades.Rol import Rol
 from Dtos.RolDTO import RolDTO
 
-from Entidades.Editorial import Editorial
-from Dtos.EditorialDTO import EditorialDTO
 
-from Entidades.Autor import Autor
-from Dtos.AutorDTO import AutorDTO
 
-from Dtos.LibroAutorDTO import LibroAutorDTO
-from Entidades.LibroAutor import LibroAutor
 
-from Entidades.Categoria import Categoria
-from Dtos.CategoriaDTO import CategoriaDTO
 
-from Entidades.LibroCategoria import LibroCategoria
-from Dtos.LibroCategoriaDTO import LibroCategoriaDTO
 
-from Entidades.Devolucion import Devolucion
-from Dtos.DevolucionDTO import DevolucionDTO
+
+
+
+
 
 
 """Mapeador Rol"""
@@ -32,85 +24,25 @@ def fila_a_rol(fila: tuple) -> Rol:
 
 """Mapeador Editorial"""
 
-def editorial_a_dto(editorial: Editorial) -> EditorialDTO:
-    """Convierte una entidad Editorial en un DTO EditorialDTO."""
-    return EditorialDTO(id=editorial.id, nombre=editorial.nombre, pais=editorial.pais)
 
-def dto_a_editorial(editorial_dto: EditorialDTO) -> Editorial:
-    """Convierte un DTO EditorialDTO en una entidad Editorial."""
-    return Editorial(id=editorial_dto.id, nombre=editorial_dto.nombre, pais=editorial_dto.pais)
-
-def fila_a_editorial(fila: tuple) -> Editorial:
-    """Convierte una fila de la base de datos (tupla) en una entidad Editorial."""
-    return Editorial(id=fila[0], nombre=fila[1], pais=fila[2])
 
 """Mapeador Autor"""
 
-def autor_a_dto(autor: Autor) -> AutorDTO:
-    return AutorDTO(id=autor.id, nombre=autor.nombre, nacionalidad=autor.nacionalidad)
 
-def dto_a_autor(autor_dto: AutorDTO) -> Autor:
-    return Autor(id=autor_dto.id, nombre=autor_dto.nombre, nacionalidad=autor_dto.nacionalidad)
-
-def fila_a_autor(fila: tuple) -> Autor:
-    return Autor(id=fila[0], nombre=fila[1], nacionalidad=fila[2])
 
 """Mapeador Categoria"""
 
-def categoria_a_dto(categoria: Categoria) -> CategoriaDTO:
-    return CategoriaDTO(id=categoria.id, nombre=categoria.nombre)
 
-def dto_a_categoria(categoria_dto: CategoriaDTO) -> Categoria:
-    return Categoria(id=categoria_dto.id, nombre=categoria_dto.nombre)
-
-def fila_a_categoria(fila: tuple) -> Categoria:
-    return Categoria(id=fila[0], nombre=fila[1])
 
 """Mapeador LibroCategoria"""
 
-def libro_categoria_a_dto(libro_categoria: LibroCategoria) -> LibroCategoriaDTO:
-    return LibroCategoriaDTO(libro_id=libro_categoria.libro_id, categoria_id=libro_categoria.categoria_id)
 
-def dto_a_libro_categoria(libro_categoria_dto: LibroCategoriaDTO) -> LibroCategoria:
-    return LibroCategoria(libro_id=libro_categoria_dto.libro_id, categoria_id=libro_categoria_dto.categoria_id)
-
-def fila_a_libro_categoria(fila: tuple) -> LibroCategoria:
-    return LibroCategoria(libro_id=fila[0], categoria_id=fila[1])
 
 """Mapeador LibroAutor"""
 
-def libro_autor_a_dto(libro_autor: LibroAutor) -> LibroAutorDTO:
-    return LibroAutorDTO(id=libro_autor.id, id_libro=libro_autor.id_libro, id_autor=libro_autor.id_autor)
 
-def dto_a_libro_autor(libro_autor_dto: LibroAutorDTO) -> LibroAutor:
-    return LibroAutor(id=libro_autor_dto.id, id_libro=libro_autor_dto.id_libro, id_autor=libro_autor_dto.id_autor)
-
-def fila_a_libro_autor(fila: tuple) -> LibroAutor:
-    return LibroAutor(id=fila[0], id_libro=fila[1], id_autor=fila[2])
 
 """Mapeador Devolucion"""
 
-def devolucion_a_dto(dev: Devolucion) -> DevolucionDTO:
-    return DevolucionDTO(
-        id=dev.id,
-        fecha_real_devolucion=dev.fecha_real_devolucion,
-        estado_libro=dev.estado_libro,
-        observaciones=dev.observaciones
-    )
 
-def dto_a_devolucion(dto: DevolucionDTO) -> Devolucion:
-    return Devolucion(
-        id=dto.id,
-        fecha_real_devolucion=dto.fecha_real_devolucion,
-        estado_libro=dto.estado_libro,
-        observaciones=dto.observaciones
-    )
-
-def fila_a_devolucion(fila: tuple) -> Devolucion:
-    return Devolucion(
-        id=fila[0],
-        fecha_real_devolucion=fila[1],
-        estado_libro=fila[2],
-        observaciones=fila[3]
-    )
